Refactor to use AyahShape (#5)
* Refactor to use AyahShape

* :bug: fix some bugs related to new ayahshap refactor (#6)

* Modified test suite to confirm to API changes

Both ayahFound and matchFound events now return AyahShape objects instead
of just surahNum, ayahNum and ayahText. The tests have been updated to
reflect this.

* Removed redundant transcription state variables.

The current surah and ayah number were stored redundantly in independent
variables as well as an AyahShape object. This commit removes those extra
variables and modifies the code to rely solely on the AyahShape object.

* Typo fix in comment