Add support for special ayat (#8)
* Added support for special ayat.

Special ayat are detected but ignored for follow along. No events
are sent for special ayat.

* Removed special ayah from test suite

Bismillah is a special ayah, and hence has no events sent for it. old
tests used Bismillah in several cases, which have now been updated to use
other non-special ayat so we get the expected events.

* Added unit and integration tests for special ayat.